* first step to using stdint.h types for portability * removed some old DJGPP code in the translator

+++ source/object.h
@@ -1,8 +1,9 @@
 #ifndef OBJECT_H_
 #define OBJECT_H_
 
+#include <stdint.h>
 
-typedef long object;
+typedef intptr_t object;
 typedef object *object_ptr;
 
 struct cleanup;
@@ -10,9 +11,9 @@ typedef struct cleanup *cleanup_ptr;
 typedef void(*cleanup_func)(object);
 
 struct cleanup {
-	long type;
+	int type;
 	union func_union{
-		long rid;
+		int rid;
 		cleanup_func builtin;
 	} func;
 	cleanup_ptr next;
@@ -20,15 +21,15 @@ struct cleanup {
 
 struct s1 {                        /* a sequence header block */
 	object_ptr base;               /* pointer to (non-existent) 0th element */
-	long length;                   /* number of elements */
-	long ref;                      /* reference count */
-	long postfill;                 /* number of post-fill objects */
+	int length;                   /* number of elements */
+	int ref;                      /* reference count */
+	int postfill;                 /* number of post-fill objects */
 	cleanup_ptr cleanup;           /* custom clean up when sequence is deallocated */
 }; /* total 20 bytes */
 
 struct d {                         /* a double precision number */
 	double dbl;                    /* double precision value */
-	long ref;                      /* reference count */
+	int ref;                      /* reference count */
 	cleanup_ptr cleanup;           /* custom clean up when sequence is deallocated */
 }; /* total 16 bytes */
